pip_requirements.txt
BeautifulSoup==3.2.0
django-endless-pagination
django-sorting==0.1
django-tagging==0.3.1
djrill==0.9.0
httplib2==0.8
openpyxl==1.6.1
python-dateutil==1.5
python-openid==2.2.5
reportlab==2.5
simplejson==3.3.1
six==1.4.1

